+++ CHANGELOG.md
@@ -4,6 +4,7 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 ## [2.2.0] - TBD
 #### Fixed
 - Correctly update `Tic-Tac-Toe` board between players
+- Set correct match type on Spectate page
 
 ## [2.1.0] - 2021-10-17
 #### Feature